# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.SystemLog:
    properties:
      client_ip:
        type: string
      created_at:
        type: string
      error_msg:
        type: string
      id:
        type: integer
      latency:
        type: string
      method:
        type: string
      path:
        type: string
      request_id:
        description: ID Unik per request
        type: string
      status_code:
        type: integer
    type: object
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      latency:
        type: string
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
host: localhost:9001
info:
  contact: {}
  description: API Server untuk manajemen Listmak
  title: Listmak Service API
  version: "1.0"
paths:
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Google callback
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Google callback
      tags:
      - auth
  /auth/google/login:
    get:
      responses:
        "307":
          description: Temporary Redirect
      summary: Redirect to Google Login
      tags:
      - auth
  /logs:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SystemLog'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all logs
      tags:
      - logs
  /logs/{request_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SystemLog'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get log by request id
      tags:
      - logs
  /users:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua user yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Membuat user baru
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create new user
      tags:
      - users
swagger: "2.0"